Preserve PostgreSQL point precision across CDC paths
Map scalar PostgreSQL point values to x/y double-precision structs and decode the native 16-byte point format during snapshot reads. Cover explicit schemas, inferred schemas, and automatic schema changes with precision-sensitive CDC tests.

Constraint: Debezium emits point coordinates as FLOAT64 while the prior RisingWave mapping used Float32
Rejected: Map point to bytea | the existing struct representation is user-facing and avoids opaque binary values
Rejected: Extend this change to point arrays | Debezium 3.2.4 does not implement POINT_ARRAY incremental schema or conversion
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep snapshot point decoding restricted to PostgreSQL POINT and exact x/y Float64 structs
Tested: cargo fmt; connector point unit tests; full RisingWave playground build; 18-module Java connector build; PostgreSQL CDC SQLLogicTest; SLT coverage check
Not-tested: PostgreSQL point arrays, which retain pre-existing behavior

+++ src/connector/src/source/cdc/external/postgres.rs
@@ -926,8 +926,8 @@ pub fn pg_type_to_rw_type(pg_type: &PgType) -> ConnectorResult<DataType> {
         PgType::TIME => DataType::Time,
         PgType::TIMETZ => DataType::Time,
         PgType::POINT => DataType::Struct(risingwave_common::types::StructType::new(vec![
-            ("x", DataType::Float32),
-            ("y", DataType::Float32),
+            ("x", DataType::Float64),
+            ("y", DataType::Float64),
         ])),
         PgType::TIMESTAMP => DataType::Timestamp,
         PgType::TIMESTAMPTZ => DataType::Timestamptz,
